feat:주관식 문제 생성

+++ src/main/java/com/example/interviewgod/domain/question/dto/response/MakeMultipleChoiceResponse.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import java.util.List;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record MakeMultipleChoiceResponse(Integer sessionId, List<Question> questionSet) {
+public record MakeMultipleChoiceResponse(Long sessionId, List<Question> questionSet) {
 
     public record Question(String question,
                               Integer question_number,
